- Changed the `ClientID` type from `int` (32-bit signed) to `long` to prevent overflows when calculating HighIDs.
- Aligned client ID management with the eMule protocol, ensuring correct identification of HighID/LowID clients.

+++ src/main/java/org/jemule/core/ClientFactory.java
@@ -31,7 +31,7 @@ import java.util.concurrent.atomic.AtomicLong;
 public class ClientFactory {
     private static final Logger log = LoggerFactory.getLogger(ClientFactory.class);
 
-    public ClientState createClient(InetAddress address, int port, int clientId) {
+    public ClientState createClient(InetAddress address, int port, long clientId) {
         if (address == null) {
             throw new IllegalArgumentException("Client address cannot be null");
         }
